+++ src/lsptoolshost/commands.ts
@@ -20,7 +20,8 @@ import {
 import TelemetryReporter from '@vscode/extension-telemetry';
 import { TelemetryEventNames } from '../shared/telemetryEventNames';
 import { registerCaptureLogsCommand } from './logging/captureLogs';
-import { registerTraceCommand } from './profiling/profiling';
+import { registerTraceCommand } from './logging/profiling';
+import { registerDumpCommand } from './logging/dump';
 import { ObservableLogOutputChannel } from './logging/observableLogOutputChannel';
 
 export function registerCommands(
@@ -91,4 +92,5 @@ function registerExtensionCommands(
     );
     registerCaptureLogsCommand(context, languageServer, outputChannel, csharpTraceChannel);
     registerTraceCommand(context, languageServer, outputChannel, csharpTraceChannel);
+    registerDumpCommand(context, languageServer, outputChannel, csharpTraceChannel);
 }